+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED
+- Fixed bug with saving/loading plugin state.
+- Made presets agnostic to tuning setting.
+
 ## [1.1.0] 2021-11-05
 - Added support for custom tunings.
 - Added controls to add noise to the kick drum sound.